--- a/.gitignore
+++ b/.gitignore
@@ -11,6 +11,7 @@
 *.iml
 *.iws
 /.tsc-cache/
+/.cache/
 /tags
 node_modules
 /npm-debug.log*