--- a/packages/calypso-build/package.json
+++ b/packages/calypso-build/package.json
@@ -52,6 +52,7 @@
 		"babel-jest": "^26.0.0",
 		"babel-loader": "^8.1.0",
 		"browserslist": "^4.8.2",
+		"cache-loader": "^4.1.0",
 		"caniuse-api": "^3.0.0",
 		"css-loader": "^3.4.2",
 		"duplicate-package-checker-webpack-plugin": "^3.0.0",