--- a/packages/calypso-build/webpack/sass.js
+++ b/packages/calypso-build/webpack/sass.js
@@ -11,13 +11,20 @@ const WebpackRTLPlugin = require( 'webpack-rtl-plugin' );
  * @param  {string[]}  _.includePaths                 Sass files lookup paths
  * @param  {string}    _.prelude                      String to prepend to each Sass file
  * @param  {object}    _.postCssConfig                PostCSS config
+ * @param  {object}    _.cacheDirectory               Directory used to store the cache
  *
  * @returns {object}                                  webpack loader object
  */
-module.exports.loader = ( { includePaths, prelude, postCssConfig = {} } ) => ( {
+module.exports.loader = ( { includePaths, prelude, postCssConfig = {}, cacheDirectory } ) => ( {
 	test: /\.(sc|sa|c)ss$/,
 	use: [
 		MiniCssExtractPluginWithRTL.loader,
+		{
+			loader: require.resolve( 'cache-loader' ),
+			options: {
+				cacheDirectory: cacheDirectory,
+			},
+		},
 		{
 			loader: require.resolve( 'css-loader' ),
 			options: {