### Background

`cache-loader` is a [Webpack loader](https://webpack.js.org/loaders/cache-loader/#root) used to cache the result of other loaders, but we are not using it!

Found by @nsakaimbo in #44696

### Changes

Introduce `cache-loader` for SassLoader. The other expensive loader (babel) already implements its own caching mechanism.

On my machine, it saves almost 1 minute when building the evergreen bundle for production. I assume there are similar savings in all webpack operations in general:

```
Command: time NODE_ENV=production CALYPSO_ENV=production yarn run build-client-evergreen

Without cache-loader (first run is with cold cache):
	1st run (cold cache): 356.33s user 31.47s system 190% cpu 3:23.74 total
	2nd run (warm cache): 214.09s user 22.98s system 140% cpu 2:48.21 total <--
	3rd run (warm cache): 210.78s user 23.25s system 142% cpu 2:43.75 total <--

With cache-loader for css (first run is with cold cache):
	1st run (cold cache): 321.52s user 26.28s system 146% cpu 3:57.37 total
	2nd run (warm cache): 146.36s user 14.46s system 139% cpu 1:54.93 total <--
	3rd run (warm cache): 144.11s user 14.23s system 141% cpu 1:51.82 total <--
```

### Testing instructions

* Run any webpack related command (`yarn build-client-evergreen`, `yarn start`...) and verify it is faster now.